fx 관련 문자열 제거, 코드 리펙토링
